refactor: Rename generic component folder
From Components to Common, to be clearer
